test-case-agent-web-app/pyproject.toml
[project]
name = "test-case-agent-web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "setuptools>=80.8.0",
    "streamlit>=1.32.2",
]
